requirements.txt
aiomysql==0.2.0
annotated-types==0.6.0
anyio==4.1.0
certifi==2023.11.17
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.101.0
graphql-core==3.2.3
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
pydantic==2.5.2
pydantic_core==2.14.5
PyMySQL==1.1.0
python-dateutil==2.8.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlmodel==0.0.14
starlette==0.27.0
strawberry-graphql==0.216.0
typing_extensions==4.9.0
uvicorn==0.23.2